Generalize getges_driver COM lookback
The upcoming update to GFS v17 will substantially change the COM paths
for the GFS package. This requires updates to the lookback capability of
getges_driver where it checks previous cycles. Rather than add additional
logical branches to handle new COM patterns, the process is generalized by
instead doing a string substitution to replace the current day and cycle
with those for the prior cycle. This method is backwards compatible and
should work for any path as long as it contains `${PDY}/${cyc}`.

Additional changes will likely be needed for GFSv17 to run prepobs stand-
alone (as in ops) to match the new `$COMINgdas` and `$COMINgfs`, but
these changes work when running prepobs as part of global workflow in
development.

Also, I have not otherwise modified the existing style of the scripts,
redefining `$COMINgdas` and `$COMINgfs`. However, this will need to be
changed before GFSv17 to stop overwriting these variables to meet NCO
requirements (see NOAA-EMC/global-workflow#293).

Refs: #17
Refs: NOAA-EMC/global-workflow#761